requirements.txt
requests>=2.31.0
openai>=1.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
# google-generativeai>=0.3.0  # если хотите использовать официальный SDK для Gemini